requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
greenlet==3.1.1
idna==3.10
loguru==0.7.2
playwright==1.49.0
pyee==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
selectolax==0.3.26
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
win32-setctime==1.1.0
